+++ hashview.py
@@ -30,6 +30,12 @@ with app.app_context():
     tasks = Tasks.query.count()
     settings = Settings.query.first()
 
+    try:
+        from authlib import jose
+    except:
+        print('\nPlease make sure that your dependencies are up to date (including installing authlib).')
+        exit(1)
+
     # If no admins exist prompt user to generate new admin account
     if users == 0:
         print('\nInitial setup detected. Hashview will now prompt you to setup an Administrative account.\n')