+++ CHANGELOG.md
@@ -7,10 +7,12 @@ Notable changes will be documented here
 - Added support for pushover & email notifications
 - Added support for data retention
 - User roles. Now you have admins and non-admins.
+- Added a last login date to the users list.
 ### Changed
 - Everything is now python
 - Moved SMTP settings to config file
 - Hashview Agent is now packaged with hashview (server) under install directory. Can be downloaded from agents menu as admin
+- Changed from itsdangerous to authlib for password reset token generation (please make sure to update your environments to include authlib).
 ### Removed
 - Removed hashview agent from local processing. If you want to run hashview AND crack hashes on the same box run the hashview-agent in a seperate screen/tmux session